### 2023-10-23 #5977 @editorialbot
[PRE REVIEW]: MiscMetabar : a R packages to facilitate visualization and Reproducibility in metabarcoding analysis
**Submitting author:** @adrientaudiere (<a href="http://orcid.org/0000-0003-1088-1182">Adrien Taudiere</a>)
**Repository:** https://github.com/adrientaudiere/MiscMetabar
**Branch with paper.md** (empty if default branch):
**Version:** v0.5
