### 2023-10-24 #5977 @Kevin-Mattheus-Moerman
@adrientaudiere I am the AEiC on this track and here to help with initial steps. I will be looking for a handling editor next. Note however we have a bit of a backlog and many editors are currently preoccupied with other submissions. Hence it may take a bit longer than usual to find a handling editor.

For the moment, could you try to address the above potentially missing DOIs? Note if some of these references are not used in the paper please remove them from the bib file. Once you've made changes you can call `@editorialbot check references` here in a comment to check the DOIs again.
