+++ paper/bibliography.bib
@@ -10,6 +10,16 @@
   publisher = {Elsevier}
 }
 
+@article{antich2021,
+  title     = {To denoise or to cluster, that is not the question: optimizing pipelines for COI metabarcoding and metaphylogeography},
+  author    = {Antich, Adri{\`a} and Palacin, Creu and Wangensteen, Owen S and Turon, Xavier},
+  journal   = {BMC bioinformatics},
+  volume    = {22},
+  pages     = {1--24},
+  year      = {2021},
+  publisher = {Springer},
+  doi       = {10.1101/2021.01.08.425760}
+}
 
 @article{Barnett2021,
   doi       = {10.21105/joss.03201},
@@ -23,6 +33,7 @@
   title     = {microViz: an R package for microbiome data visualization and statistics},
   journal   = {Journal of Open Source Software}
 }
+
 @article{bolyen2019,
   doi       = {10.1038/s41587-019-0209-9},
   title     = {Reproducible, interactive, scalable and extensible microbiome data science using QIIME 2},
@@ -66,6 +77,18 @@
   doi    = {10.18129/B9.bioc.mia}
 }
 
+@article{forster2019,
+  doi       = {10.1111/1462-2920.14764},
+  title     = {Improving eDNA-based protist diversity assessments using networks of amplicon sequence variants},
+  author    = {Forster, Dominik and Lentendu, Guillaume and Filker, Sabine and Dubois, Elyssa and Wilding, Thomas A and Stoeck, Thorsten},
+  journal   = {Environmental microbiology},
+  volume    = {21},
+  number    = {11},
+  pages     = {4109--4124},
+  year      = {2019},
+  publisher = {Wiley Online Library}
+}
+
 @article{froslev2017,
   doi       = {10.1038/s41467-017-01312-x},
   title     = {Algorithm for post-clustering curation of DNA amplicon data yields reliable biodiversity estimates},
@@ -76,6 +99,20 @@
   pages     = {1188},
   year      = {2017},
   publisher = {Nature Publishing Group UK London}
+}
+
+@article{Huang2021,
+  author    = {Huang, Ruizhu and Soneson, Charlotte and Ernst, Felix and Rue-Albrecht, Kevin and Yu, Guangchuang and Hicks, Stephanie and Robinson, Mark},
+  journal   = {F1000Research},
+  title     = {TreeSummarizedExperiment: a S4 class for data with hierarchical structure},
+  year      = {2021},
+  month     = {3},
+  pages     = {1246},
+  volume    = {9},
+  date      = {2021-03-02},
+  day       = {2},
+  doi       = {10.12688/f1000research.26669.2},
+  publisher = {F1000 Research Ltd}
 }
 
 @article{jalili2020,
@@ -89,6 +126,28 @@
   publisher = {Oxford University Press}
 }
 
+@article{joos2020,
+  doi       = {10.1186/s12864-020-07126-4},
+  title     = {Daring to be differential: metabarcoding analysis of soil and plant-related microbial communities using amplicon sequence variants and operational taxonomical units},
+  author    = {Joos, Lisa and Beirinckx, Stien and Haegeman, Annelies and Debode, Jane and Vandecasteele, Bart and Baeyen, Steve and Goormachtig, Sofie and Clement, Lieven and De Tender, Caroline},
+  journal   = {BMC genomics},
+  volume    = {21},
+  number    = {1},
+  pages     = {1--17},
+  year      = {2020},
+  publisher = {BioMed Central}
+}
+
+@article{kauserud2023,
+  doi       = {10.1016/j.funeco.2023.101274},
+  title     = {ITS alchemy: On the use of ITS as a DNA marker in fungal ecology},
[…]
