+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   beta-div: beta-div.html
   filter: filter.html
   import_export_track: import_export_track.html
-last_built: 2023-10-23T15:47Z
+last_built: 2023-10-25T07:19Z
 urls:
   reference: https://adrientaudiere.github.io/MiscMetabar/reference
   article: https://adrientaudiere.github.io/MiscMetabar/articles
