### 2023-10-25 8fb8d6c
Improve vignettes and README
