### 2023-10-25 #5977 @adrientaudiere
Hi @Kevin-Mattheus-Moerman,
Thanks for the rapid feedback. I cleaned the references and added valid DOI. I also improved the README file with additional information, it was obviously necessary!
