### 2023-11-08 #6038 @editorialbot
[REVIEW]: MiscMetabar: an R package to facilitate visualization and Reproducibility in metabarcoding analysis
**Submitting author:** @adrientaudiere (<a href="http://orcid.org/0000-0003-1088-1182">Adrien Taudiere</a>)
**Repository:** https://github.com/adrientaudiere/MiscMetabar
**Branch with paper.md** (empty if default branch):
**Version:** v0.52
