+++ R/beta_div_test.R
@@ -49,21 +49,23 @@ graph_test_pq <- function(physeq,
   }
 
   if (na_remove) {
-    new_physeq <- subset_samples_pq(physeq, !is.na(physeq@sam_data[[fact]]))
+    new_physeq <-
+      subset_samples_pq(physeq,!is.na(physeq@sam_data[[fact]]))
     if (nsamples(physeq) - nsamples(new_physeq) > 0) {
-      message(paste0(
-        nsamples(physeq) - nsamples(new_physeq),
-        " were discarded due to NA in variables present in formula."
-      ))
+      message(
+        paste0(
+          nsamples(physeq) - nsamples(new_physeq),
+          " were discarded due to NA in variables present in formula."
+        )
+      )
     }
     physeq <- new_physeq
   }
 
   res_graph_test <- phyloseqGraphTest::graph_perm_test(physeq,
-    sampletype = fact,
-    nperm = nperm,
-    ...
-  )
+                                                       sampletype = fact,
+                                                       nperm = nperm,
+                                                       ...)
   if (!return_plot) {
     return(res_graph_test)
   } else {
@@ -71,8 +73,11 @@ graph_test_pq <- function(physeq,
       labs(
         title = title,
         subtitle = paste(
-          "pvalue = ", res_graph_test$pval,
-          "(", length(res_graph_test$perm), " permutations)"
+          "pvalue = ",
+          res_graph_test$pval,
+          "(",
+          length(res_graph_test$perm),
+          " permutations)"
         )
       )
     return(p)
@@ -118,15 +123,14 @@ graph_test_pq <- function(physeq,
 #' @export
 #' @author Adrien Taudière
 
-adonis_pq <- function(
-    physeq,
-    formula,
-    dist_method = "bray",
-    merge_sample_by = NULL,
-    na_remove = FALSE,
-    correction_for_sample_size = FALSE,
-    rarefy_nb_seqs = FALSE,
-    ...) {
+adonis_pq <- function(physeq,
+                      formula,
+                      dist_method = "bray",
+                      merge_sample_by = NULL,
+                      na_remove = FALSE,
+                      correction_for_sample_size = FALSE,
+                      rarefy_nb_seqs = FALSE,
+                      ...) {
   physeq <- clean_pq(
     physeq,
     force_taxa_as_columns = TRUE,
@@ -137,9 +141,13 @@ adonis_pq <- function(
   )
 
   if (dist_method %in% c("aitchison", "robust.aitchison")) {
-    phy_dist <- paste0('vegan::vegdist(as.matrix(physeq@otu_table), method="', dist_method, '")')
+    phy_dist <-
+      paste0('vegan::vegdist(as.matrix(physeq@otu_table), method="',
+             dist_method,
+             '")')
   } else {
-    phy_dist <- paste0('phyloseq:::distance(physeq, method="', dist_method, '")')
+    phy_dist <-
+      paste0('phyloseq:::distance(physeq, method="', dist_method, '")')
   }
 
   .formula <- stats::reformulate(formula, response = phy_dist)
@@ -152,13 +160,16 @@ adonis_pq <- function(
     new_physeq <- physeq
     for (tl in term_lab) {
       print(tl)
-      new_physeq <- subset_samples_pq(new_physeq, !is.na(physeq@sam_data[[tl]]))
+      new_physeq <-
+        subset_samples_pq(new_physeq,!is.na(physeq@sam_data[[tl]]))
     }
     if (nsamples(physeq) - nsamples(new_physeq) > 0) {
-      message(paste0(
[…]
