+++ vignettes/states_of_fields_in_R.Rmd
@@ -2,7 +2,7 @@
 title: "Metabarcoding with R"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Reclustering}
+  %\VignetteIndexEntry{Metabarcoding with R}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 bibliography: ../paper/bibliography.bib
