+++ _pkgdown.yml
@@ -13,16 +13,16 @@ navbar:
     text: Articles
     menu:
     - text: Tutorials
-    - text: Reclustering
-      href: articles/Reclustering.html
     - text: Alpha-diversity
       href: articles/alpha-div.html
     - text: Beta-diversity
       href: articles/beta-div.html
-    - text: Import/export and track 
+    - text: Import/export and track
       href: articles/import_export_track.html
     - text: Filter taxa and samples
       href: articles/filter.html
+    - text: Reclustering
+      href: articles/Reclustering.html
     - text: -------
     - text: Tutorials with published dataset
     - text: Tengeler
@@ -36,6 +36,6 @@ navbar:
     - text: Rules of code
       href: articles/Rules.html
 
-      
+
 development:
   mode: auto
