### 2023-11-09 37c0671
Minor corrections to improve codefactor.io
