### 2023-11-09 8f1a57e
Create CONTRIBUTING.md
