+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   filter: filter.html
   import_export_track: import_export_track.html
   states_of_fields_in_R: states_of_fields_in_R.html
-last_built: 2023-11-09T16:12Z
+last_built: 2023-11-09T20:01Z
 urls:
   reference: https://adrientaudiere.github.io/MiscMetabar/reference
   article: https://adrientaudiere.github.io/MiscMetabar/articles
