### 2023-11-10 7ca9f19
Feature : new function tbl_sum_samdata
