+++ DESCRIPTION
@@ -30,6 +30,7 @@ Suggests:
     ggh4x,
     ggridges,
     ggVennDiagram,
+    gtsummary,
     grDevices,
     grid,
     gridExtra,
