### 2023-12-01 9169a72
feat: add function tbl_sum_samdata
