+++ man/multipatt_pq.Rd
@@ -1,0 +1,48 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/beta_div_test.R
+\name{multipatt_pq}
+\alias{multipatt_pq}
+\title{Test and plot multipatt result}
+\usage{
+multipatt_pq(
+  physeq,
+  fact,
+  p_adjust_method = "BH",
+  pval = 0.05,
+  control = how(nperm = 999),
+  ...
+)
+}
+\arguments{
+\item{physeq}{(required): a \code{\link{phyloseq-class}} object obtained
+using the \code{phyloseq} package.}
+
+\item{fact}{(required) Name of the factor in \code{physeq@sam_data} used to plot
+different lines}
+
+\item{p_adjust_method}{(chr, default "BH"): the method used to adjust p-value}
+
+\item{pval}{(int, default 0.05): the value to determine the significance of
+LCBD}
+
+\item{control}{see \code{?indicspecies::multipatt()}}
+
+\item{...}{Others arguments passed on to \code{\link[indicspecies:multipatt]{indicspecies::multipatt()}} function}
+}
+\value{
+A ggplot object
+}
+\description{
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
+
+A wrapper for the \code{\link[indicspecies:multipatt]{indicspecies::multipatt()}} function in the case of \code{physeq}
+object.
+}
+\examples{
+multipatt_pq(subset_samples(data_fungi, !is.na(Time)), fact="Time")
+multipatt_pq(subset_samples(data_fungi, !is.na(Time)), fact="Time",
+   max.order = 1, control = how(nperm=9999))
+}
+\author{
+Adrien Taudière
+}
