### 2023-12-01 3e1fcce
fix : fix a bug in compare_pairs_pq
