### 2023-12-05 8db68b9
Feat: add mumu_pq()
