+++ R/dada_phyloseq.R
@@ -111,9 +111,11 @@ clean_pq <- function(physeq,
   }
 
   if (sum(grepl("^0", sample_names(physeq)) > 0) && !silent) {
-    message("At least one sample name start with a zero.
+    message(
+      "At least one sample name start with a zero.
     That can be a problem for some phyloseq functions such as
-    plot_bar and psmelt.")
+    plot_bar and psmelt."
+    )
   }
 
   if (force_taxa_as_columns && force_taxa_as_rows) {
@@ -121,16 +123,20 @@ clean_pq <- function(physeq,
   }
 
   if (force_taxa_as_columns && taxa_are_rows(physeq)) {
-    otu_table(physeq) <- otu_table(t(as.matrix(unclass(physeq@otu_table))),
-      taxa_are_rows = FALSE
-    )
+    otu_table(physeq) <-
+      otu_table(t(as.matrix(unclass(
+        physeq@otu_table
+      ))),
+      taxa_are_rows = FALSE)
     message("Taxa are now in columns.")
   }
 
   if (force_taxa_as_rows && !taxa_are_rows(physeq)) {
-    otu_table(physeq) <- otu_table(t(as.matrix(unclass(physeq@otu_table))),
-      taxa_are_rows = TRUE
-    )
+    otu_table(physeq) <-
+      otu_table(t(as.matrix(unclass(
+        physeq@otu_table
+      ))),
+      taxa_are_rows = TRUE)
     message("Taxa are now in rows.")
   }
 
@@ -152,18 +158,29 @@ clean_pq <- function(physeq,
   }
 
   if (verbose) {
-    message(paste(
-      "Cleaning suppress", ntaxa(physeq) - ntaxa(new_physeq), "taxa (",
-      paste(taxa_names(physeq)[taxa_sums(physeq) == 0], collapse = " / "), ") and",
-      nsamples(physeq) - nsamples(new_physeq),
-      "sample(s) (", paste(sample_names(physeq)[sample_sums(physeq) == 0], collapse = " / "), ")."
-    ))
+    message(
+      paste(
+        "Cleaning suppress",
+        ntaxa(physeq) - ntaxa(new_physeq),
+        "taxa (",
+        paste(taxa_names(physeq)[taxa_sums(physeq) == 0], collapse = " / "),
+        ") and",
+        nsamples(physeq) - nsamples(new_physeq),
+        "sample(s) (",
+        paste(sample_names(physeq)[sample_sums(physeq) == 0], collapse = " / "),
+        ")."
+      )
+    )
   } else if (!silent) {
-    message(paste(
-      "Cleaning suppress", ntaxa(physeq) - ntaxa(new_physeq), "taxa and",
-      nsamples(physeq) - nsamples(new_physeq),
-      "samples."
-    ))
+    message(
+      paste(
+        "Cleaning suppress",
+        ntaxa(physeq) - ntaxa(new_physeq),
+        "taxa and",
+        nsamples(physeq) - nsamples(new_physeq),
+        "samples."
+      )
+    )
   }
 
   verify_pq(new_physeq)
@@ -205,12 +222,11 @@ clean_pq <- function(physeq,
 #'   each object.
 #' @export
 
-track_wkflow <- function(
-    list_of_objects,
-    obj_names = NULL,
-    clean_pq = FALSE,
-    taxonomy_rank = NULL,
-    ...) {
+track_wkflow <- function(list_of_objects,
+                         obj_names = NULL,
+                         clean_pq = FALSE,
+                         taxonomy_rank = NULL,
+                         ...) {
   message("Compute the number of sequences")
   if (!is.null(obj_names)) {
     names(list_of_objects) <- obj_names
[…]
