### 2023-12-05 2c44849
add details about citation for wrapper functions
