### 2023-12-06 adrientaudiere/MiscMetabar#37 @tkchafin
[JOSS REVIEW] Minor manuscript comments 
Minor comments on MS
- L9: Should this be regions (not singular)?
- L20: plateforms -> platforms
- L38: smith2019? (citation formatting)
- L47: "...allows the novice..." might be better reworded as "This plot provides an information-rich structural overview of the phyloseq object" or similar. Also sentence missing period
- L55: Function name rendering off-page
L63: Pipeline -> pipelines?

- Table 1: This is redundantly labeled as "Figure 1: Table 1: " also some overlapping text in the table, maybe check this is being rendered correctly
- Figure 1: Labelled wrong as "Figure 2: Figure 1: ..."
