### 2023-12-06 adrientaudiere/MiscMetabar#38 @tkchafin
[JOSS REVIEW] Installation instructions 
Are the installation instructions correct? I can't get installation to work without first manually installing dada2 and phyloseq

```
if (!require("BiocManager", quietly = TRUE))
  install.packages("BiocManager")

BiocManager::install("dada2")
BiocManager::install("phyloseq")
```
