### 2023-12-06 adrientaudiere/MiscMetabar#39 @tkchafin
[JOSS REVIEW] Missing dependency ggVennDiagram
given_pq() function doesn't work following install
```
Error in loadNamespace(x): there is no package called 'ggVennDiagram'
```

This is easily resolved by manually installing ggVennDiagram, but you may need to take another look at the requirements for the package to make sure all dependencies are covered.
