### 2023-12-06 #6038 @tkchafin
@Kevin-Mattheus-Moerman  I've completed my review. I have posted some minor issues regarding the [installation](https://github.com/adrientaudiere/MiscMetabar/issues/38) instructions and [missing dependencies](https://github.com/adrientaudiere/MiscMetabar/issues/39), which I expect will be easy to resolve. I've also posted some [minor editorial notes](https://github.com/adrientaudiere/MiscMetabar/issues/37) for the paper (typos and small grammatical corrections). Once these are resolved I can check off the final item in my checklist!
