+++ paper/paper.bib
@@ -1,4 +1,4 @@
-﻿@article{altschul1990,
+@article{altschul1990,
   doi       = {10.32388/rhq6vj},
   title     = {Basic local alignment search tool},
   author    = {Altschul, Stephen F and Gish, Warren and Miller, Webb and Myers, Eugene W and Lipman, David J},
@@ -101,6 +101,14 @@
   url     = {https://doi.org/10.21105/joss.02959}
 }
 
+@software{mahe2023,
+  url     = {https://github.com/frederic-mahe/mumu/},
+  year    = {2023},
+  version = {1.0.2},
+  author  = {Frédéric Mahé},
+  title   = {mumu: post-clustering curation tool for metabarcoding data}
+}
+
 @article{McCauley2022,
   author   = {McCauley, Mark and Goulet, Tamar and Jackson, Colin and Loesgen, Sandra},
   journal  = {Research Square},
@@ -133,7 +141,6 @@
   doi       = {10.5281/zenodo.3923184},
   url       = {https://doi.org/10.5281/zenodo.3923184}
 }
-
 @article{mcmurdie2013,
   doi     = {10.1371/journal.pone.0061217},
   author  = {Paul J. McMurdie and Susan Holmes},
@@ -145,7 +152,8 @@
   year    = {2013},
   url     = {http://dx.plos.org/10.1371/journal.pone.0061217}
 }
-@article{ondov2011,
+
+ @article{ondov2011,
   title     = {Interactive metagenomic visualization in a Web browser},
   author    = {Ondov, Brian D and Bergman, Nicholas H and Phillippy, Adam M},
   journal   = {BMC bioinformatics},
@@ -156,7 +164,7 @@
   publisher = {BioMed Central}
 }
 
- @article{pauvert2019,
+@article{pauvert2019,
   doi       = {10.1016/j.funeco.2019.03.005},
   title     = {Bioinformatics matters: The accuracy of plant and soil fungal community data is highly dependent on the metabarcoding pipeline},
   author    = {Pauvert, Charlie and Bu{\'e}e, Marc and Laval, Val{\'e}rie and Edel-Hermann, V{\'e}ronique and Fauchery, Laure and Gautier, Ang{\'e}lique and Lesur, Isabelle and Vallance, Jessica and Vacher, Corinne},
@@ -203,7 +211,7 @@
   publisher = {Am Soc Microbiol}
 }
 
-@article{Smith2019,
+@article{smith2019,
   doi       = {10.21105/joss.01442},
   url       = {https://doi.org/10.21105/joss.01442},
   year      = {2019},
