### 2023-12-06 d78ac63
feat: function install_pkg_needed
