### 2023-12-06 adrientaudiere/MiscMetabar#39 @adrientaudiere
Oh yes, this is because I have many suggested packages. The package `ggVennDiagram` is only used in one (non-central) function of the package. So, it is not installed with MiscMetabar. There are two solutions :
- Let the user deal with that by installing the package (the error text is clear enough). This is the solution adopted by many packages, including phyloseq.
-  Install automatically the package if needed by the function. I plan to adopt this solution in the future to facilitate user life. I already coded a function `install_pkg_needed` in the dev branch of MiscMetabar, but I require some time to implement this solution for each case.

I hope I have answered your remark.
