### 2023-12-06 adrientaudiere/MiscMetabar#37 @adrientaudiere
Thank you very much for your review. I corrected my paper using all of your comments.
Best,
Adrien
