### 2023-12-06 adrientaudiere/MiscMetabar#39 @tkchafin
Ok, as long as you are aware of the issue -- I only brought it up because this was one of the example functions in the README. You might consider adding a note in the README for that function (e.g., install_packages("ggVennDiagram") library(ggVennDiagram) before the function is called
