### 2023-12-06 bbaf87e
Update installation script
