### 2023-12-08 adrientaudiere/MiscMetabar#42 @iimog
[JOSS REVIEW] Minor manuscript comments
Minor comments:
- title: "a R packages" → "an R package"
- title: "Reproducibility" → "reproducibility" (or consistent title case)
- l. 50 ASV is used for the first time, but only introduced in l. 64
- Figure 1: The font in the table does not render correctly in my pdf (both on Windows and Linux), nor in the markdown (the transparency further impacts readability on the dark mode, see screenshots below). I suggest, adding another version of the svg with white background and font to path.

Screenshot from pdf:
![screenshot from pdf](https://github.com/adrientaudiere/MiscMetabar/assets/7403236/4df6c192-fa4d-4741-9008-3971c3eefdde)
Screenshot from rendered markdown (dark mode):
![image](https://github.com/adrientaudiere/MiscMetabar/assets/7403236/2c8ac5c9-1929-4344-9a60-c86225457fe6)

Related to https://github.com/openjournals/joss-reviews/issues/6038
