### 2023-12-08 b6f2c29
feat: add 3 functions
