+++ R/dada_phyloseq.R
@@ -422,8 +422,6 @@ track_wkflow <- function(list_of_objects,
 #' @return A list of dataframe. cf [track_wkflow()] for more information
 #'
 #' @export
-#' @md
-#'
 #' @author Adrien Taudière
 #'
 #' @examples
@@ -1393,22 +1391,22 @@ mumu_pq <- function(physeq,
       unlink("temp.uc")
     }
 
-  if (file.exists("log.txt")) {
-    unlink("temp.uc")
+    if (file.exists("log.txt")) {
+      unlink("temp.uc")
+    }
+    if (file.exists("match_list.txt")) {
+      unlink("match_list.txt")
+    }
+    if (file.exists("otu_table.csv")) {
+      unlink("otu_table.csv")
+    }
+    if (file.exists("new_OTU.tablemumu")) {
+      unlink("new_OTU.tablemumu")
+    }
   }
-  if (file.exists("match_list.txt")) {
-    unlink("match_list.txt")
-  }
-  if (file.exists("otu_table.csv")) {
-    unlink("otu_table.csv")
-  }
-  if (file.exists("new_OTU.tablemumu")) {
-    unlink("new_OTU.tablemumu")
-  }
-}
 
-return(list("new_physeq" = new_physeq,
-            "mumu_results" = result_mumu))
+  return(list("new_physeq" = new_physeq,
+              "mumu_results" = result_mumu))
 }
 ################################################################################
 
@@ -1546,7 +1544,8 @@ subset_taxa_pq <- function(physeq,
   cond <- condition[match(taxa_names(new_physeq), names(condition))]
   cond[is.na(cond)] <- FALSE
 
-  old_MA <- as(otu_table(new_physeq, taxa_are_rows = TRUE), "matrix")
+  old_MA <-
+    as(otu_table(new_physeq, taxa_are_rows = TRUE), "matrix")
   new_MA <- old_MA[cond,]
 
   if (!is.matrix(new_MA)) {
@@ -1597,7 +1596,6 @@ subset_taxa_pq <- function(physeq,
 #' @return a physeq object with one sample
 #'
 #' @export
-#' @md
 #'
 #' @author Adrien Taudière
 #'
@@ -1662,8 +1660,6 @@ select_one_sample <- function(physeq, sam_name, silent = FALSE) {
 #' @return a physeq object with a larger slot tax_table
 #'
 #' @export
-#' @md
-#'
 #' @examples
 #' # example code
 #'
@@ -1727,3 +1723,332 @@ tbl_sum_samdata <- function(physeq, remove_col_unique_value = TRUE, ...) {
   return(tbl_sum)
 }
 ################################################################################
+
+
+#' Add information about Guild for FUNGI using [FUNGuildR::funguild_assign()]
+#'
+#' @description
+#' `r lifecycle::badge("experimental")`
+#'
+#' @inheritParams clean_pq
+#' @param taxLevels Name of the 7 columns in tax_table required by funguild
+#'
+#' @return A new object of class `physeq` with Guild information added to
+#'   `tax_table` slot
+#' @export
+#' @author Adrien Taudière
+#' @examples
+#' df <- subset_taxa_pq(data_fungi, taxa_sums(data_fungi)>5000)
+#' df <- add_funguild_info(df,
+#'   taxLevels = c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species"))
+#' sort(table(df@tax_table[,"guild"]), decreasing = TRUE)
+#' @details
+#' This function is mainly a wrapper of the work of others.
+#'   Please make a reference to `FUNGuildR::funguild_assign()` if you
+#'   use this function.
+#' @seealso [plot_guild_pq()]
+
+add_funguild_info <- function(physeq,
+                              taxLevels = c("Kingdom",
[…]
