+++ R/beta_div_test.R
@@ -31,7 +31,7 @@
 #' @return A \code{\link{ggplot}}2 plot with a subtitle indicating the pvalue
 #' and the number of permutations
 #' @details
-#' This function is mainly a wrapper of the work of others. 
+#' This function is mainly a wrapper of the work of others.
 #'   Please cite `phyloseqGraphTest` package.
 #' @export
 
@@ -52,7 +52,7 @@ graph_test_pq <- function(physeq,
 
   if (na_remove) {
     new_physeq <-
-      subset_samples_pq(physeq,!is.na(physeq@sam_data[[fact]]))
+      subset_samples_pq(physeq, !is.na(physeq@sam_data[[fact]]))
     if (nsamples(physeq) - nsamples(new_physeq) > 0) {
       message(
         paste0(
@@ -125,8 +125,8 @@ graph_test_pq <- function(physeq,
 #' @export
 #' @author Adrien Taudière
 #' @details
-#' This function is mainly a wrapper of the work of others. 
-#'   Please make a reference to `vegan::adonis2()` if you 
+#' This function is mainly a wrapper of the work of others.
+#'   Please make a reference to `vegan::adonis2()` if you
 #'   use this function.
 
 adonis_pq <- function(physeq,
@@ -167,7 +167,7 @@ adonis_pq <- function(physeq,
     for (tl in term_lab) {
       print(tl)
       new_physeq <-
-        subset_samples_pq(new_physeq,!is.na(physeq@sam_data[[tl]]))
+        subset_samples_pq(new_physeq, !is.na(physeq@sam_data[[tl]]))
     }
     if (nsamples(physeq) - nsamples(new_physeq) > 0) {
       message(
@@ -224,8 +224,8 @@ adonis_pq <- function(physeq,
 #' LCBD_pq(data_fungi, nperm=100)
 #' LCBD_pq(data_fungi, nperm=100, method = "jaccard")
 #' @author Adrien Taudière
-#' This function is mainly a wrapper of the work of others. 
-#'   Please make a reference to `despatial::beta.div()` if you 
+#' This function is mainly a wrapper of the work of others.
+#'   Please make a reference to `despatial::beta.div()` if you
 #'   use this function.
 LCBD_pq <- function(physeq,
                     p_adjust_method =  "BH",
@@ -281,8 +281,8 @@ LCBD_pq <- function(physeq,
 #'         axis.title.x = element_text(size=6))
 #' @author Adrien Taudière
 #' @details
-#' This function is mainly a wrapper of the work of others. 
-#'   Please make a reference to `vegan::beta.div()` if you 
+#' This function is mainly a wrapper of the work of others.
+#'   Please make a reference to `vegan::beta.div()` if you
 #'   use this function.
 plot_LCBD_pq <- function(physeq,
                          p_adjust_method =  "BH",
@@ -301,7 +301,7 @@ plot_LCBD_pq <- function(physeq,
   ))) > 0) {
     warning("At least one sample was removed by the beta.div function")
     sam_data <-
-      sam_data[rownames(sam_data) %in% names(resBeta$LCBD), ]
+      sam_data[rownames(sam_data) %in% names(resBeta$LCBD),]
   }
 
   resLCBD <- tibble(
@@ -313,7 +313,7 @@ plot_LCBD_pq <- function(physeq,
 
   if (only_plot_significant) {
     p_LCBD <-
-      ggplot(filter(resLCBD, p.adj < pval), aes(x = LCBD, y = reorder(Sample_names,-LCBD, sum))) +
+      ggplot(filter(resLCBD, p.adj < pval), aes(x = LCBD, y = reorder(Sample_names, -LCBD, sum))) +
       geom_point(size = 4)
     if (is.null(sam_variables)) {
       return(p_LCBD)
[…]
