+++ man/plot_deseq2_pq.Rd
@@ -64,6 +64,9 @@ A \code{\link{ggplot}}2 plot representing DESeq2 results
 \description{
 \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
 }
+\details{
+Please cite \code{DESeq2} package if you use chis function.
+}
 \examples{
 data("GlobalPatterns")
 GP <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == "Archaea")
