### 2023-12-08 adrientaudiere/MiscMetabar#44 @iimog
[JOSS REVIEW] Two README files
There are two README files in the repository: `README.md` and `README.Rmd`.
These two files are out of sync (e.g. the former has instructions on manually installing Bioc packages, the latter doesn't). `README.md` seems to be rendered by default on GitHub while the `README.Rmd` seems to be used for the pkgdown site.
I suggest to use just a single README file to avoid this problem.

There is also a small typo in the README ("pkdown" → "pkgdown").

Related to https://github.com/openjournals/joss-reviews/issues/6038
