### 2023-12-08 adrientaudiere/MiscMetabar#45 @iimog
[JOSS REVIEW] Link to Tengeler vignette broken
When clicking the link to the Tengeler article on the pkgdown site: https://adrientaudiere.github.io/MiscMetabar/articles/tengeler.html
A 404 (page not found) message is shown.

Related to https://github.com/openjournals/joss-reviews/issues/6038
