### 2023-12-08 adrientaudiere/MiscMetabar#45 @adrientaudiere
Oh, good point. I missed the point that quarto format (.qmd) is not supported by pkgdown.
