### 2023-12-08 adrientaudiere/MiscMetabar#44 @adrientaudiere
The use of two README is required by pkgdowm (thank for the typo). But you are true, in the master branch, README.md and README.Rmd are out of sync. It will be solved in version 0.51 of MiscMetabar.
