### 2023-12-08 adrientaudiere/MiscMetabar#42 @adrientaudiere
Thanks for the suggestion. This problem is also true (but least visible) in figure 2. I will change .svg by .png images to bypass this problem in the next version of the paper.md.
