+++ DESCRIPTION
@@ -81,8 +81,8 @@ Remotes:
     github::adrientaudiere/lulu,
     github::david-barnett/microViz,
     github::erocoar/gghalves,
-    bioc::release/phyloseq,
-    bioc::release/dada2
+    bioc::phyloseq,
+    bioc::dada2
 Config/testthat/edition: 3
 Config/testthat/parallel: true
 VignetteBuilder: knitr
