### 2023-12-08 adrientaudiere/MiscMetabar#42 @adrientaudiere
> title: "a R packages" → "an R package"

Correction done

> title: "Reproducibility" → "reproducibility" (or consistent title case),

Correction done

> l. 50 ASV is used for the first time, but only introduced in l. 64

You are completely right. I reverse the _Post-clustering_ and the _Cleaning and filtering_ sections.
