### 2023-12-08 adrientaudiere/MiscMetabar#48 @iimog
[JOSS REVIEW] Smartphone section in the bug report issue template
The bug report issue template does not seem to fit for this purpose. I found the **Smartphone** section to be surprising and irrelevant. Also for desktop it might be more meaningful for you, to ask for the output of `sessionInfo()` rather than the Browser. Below is the current bug report template for reference.

Related to https://github.com/openjournals/joss-reviews/issues/6038

**Describe the bug**
A clear and concise description of what the bug is.

**To Reproduce**
Steps to reproduce the behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See error

**Expected behavior**
A clear and concise description of what you expected to happen.

**Screenshots**
If applicable, add screenshots to help explain your problem.

 - OS: [e.g. iOS]
 - Browser [e.g. chrome, safari]
 - Version [e.g. 22]

**Smartphone (please complete the following information):**
 - Device: [e.g. iPhone6]
 - OS: [e.g. iOS8.1]
 - Browser [e.g. stock browser, safari]
 - Version [e.g. 22]

**Additional context**
Add any other context about the problem here.
