+++ paper/paper.bib
@@ -26,7 +26,7 @@
 @article{bolyen2019,
   doi       = {10.1038/s41587-019-0209-9},
   title     = {Reproducible, interactive, scalable and extensible microbiome data science using QIIME 2},
-  author    = {Bolyen, Evan and Rideout, Jai Ram and Dillon, Matthew R and Bokulich, Nicholas A and Abnet, Christian C and Al-Ghalith, Gabriel A and Alexander, Harriet and Alm, Eric J and Arumugam, Manimozhiyan and Asnicar, Francesco and others},
+  author    = {Bolyen, Evan and Rideout, Jai Ram and Dillon, MaThew R and Bokulich, Nicholas A and Abnet, Christian C and Al-Ghalith, Gabriel A and Alexander, Harriet and Alm, Eric J and Arumugam, Manimozhiyan and Asnicar, Francesco and others},
   journal   = {Nature biotechnology},
   volume    = {37},
   number    = {8},
@@ -89,7 +89,7 @@
   publisher = {Oxford University Press}
 }
 
-@article{Landau2021,
+@article{landau2021,
   doi     = {10.21105/joss.02959},
   title   = {The targets R package: a dynamic Make-like function-oriented pipeline toolkit for reproducibility and high-performance computing},
   author  = {William Michael Landau},
