+++ R/dada_phyloseq.R
@@ -124,19 +124,23 @@ clean_pq <- function(physeq,
 
   if (force_taxa_as_columns && taxa_are_rows(physeq)) {
     otu_table(physeq) <-
-      otu_table(t(as.matrix(unclass(
-        physeq@otu_table
-      ))),
-      taxa_are_rows = FALSE)
+      otu_table(
+        t(as.matrix(unclass(
+          physeq@otu_table
+        ))),
+        taxa_are_rows = FALSE
+      )
     message("Taxa are now in columns.")
   }
 
   if (force_taxa_as_rows && !taxa_are_rows(physeq)) {
     otu_table(physeq) <-
-      otu_table(t(as.matrix(unclass(
-        physeq@otu_table
-      ))),
-      taxa_are_rows = TRUE)
+      otu_table(
+        t(as.matrix(unclass(
+          physeq@otu_table
+        ))),
+        taxa_are_rows = TRUE
+      )
     message("Taxa are now in rows.")
   }
 
@@ -248,17 +252,19 @@ track_wkflow <- function(list_of_objects,
       } else if (inherits(object, "matrix")) {
         sum(object, na.rm = TRUE)
       } else if (is.character(object[1]) &&
-                 length(object[1]) == 1 &&
-                 file.exists(object[1])) {
+        length(object[1]) == 1 &&
+        file.exists(object[1])) {
         if (summary(file(object[[1]]))$class == "gzfile") {
           pbapply::pbsapply(object, function(x) {
             as.numeric(system(paste("zcat ", x, " | grep -c '^+$'", sep = ""),
-                              intern = TRUE))
+              intern = TRUE
+            ))
           })
         } else if (grepl("\\.fastq$", object[1])) {
           pbapply::pbsapply(object, function(x) {
             as.numeric(system(paste("cat ", x, " | grep -c '^+$'", sep = ""),
-                              intern = TRUE))
+              intern = TRUE
+            ))
           })
         } else {
           stop("Files must be either gzfile or .fastq")
@@ -289,7 +295,7 @@ track_wkflow <- function(list_of_objects,
       } else if (inherits(object[[1]], "dada")) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[2]
       } else if (is.data.frame(object[[1]]) &&
-                 all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
+        all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[2]
       } else if (inherits(object, "derep")) {
         length(unique(names(object$uniques)))
@@ -315,15 +321,15 @@ track_wkflow <- function(list_of_objects,
       } else if (inherits(object[[1]], "dada")) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[1]
       } else if (is.data.frame(object[[1]]) &&
-                 all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
+        all(c("sequence", "abundance") %in% colnames(object[[1]]))) {
         dim(suppressMessages(dada2::makeSequenceTable(object)))[1]
       } else if (inherits(object, "derep")) {
         1
       } else if (inherits(object[[1]], "derep")) {
         length(object)
       } else if (is.character(object[1]) &&
-                 length(object[1]) == 1 &&
-                 file.exists(object[1])) {
+        length(object[1]) == 1 &&
+        file.exists(object[1])) {
         length(object)
       } else {
         NA
@@ -337,7 +343,7 @@ track_wkflow <- function(list_of_objects,
[…]

[…]
