+++ vignettes/bibliography.bib
@@ -1,0 +1,347 @@
+﻿@article{altschul1990,
+  doi       = {10.32388/rhq6vj},
+  title     = {Basic local alignment search tool},
+  author    = {Altschul, Stephen F and Gish, Warren and Miller, Webb and Myers, Eugene W and Lipman, David J},
+  journal   = {Journal of molecular biology},
+  volume    = {215},
+  number    = {3},
+  pages     = {403--410},
+  year      = {1990},
+  publisher = {Elsevier}
+}
+
+
+@article{Barnett2021,
+  doi       = {10.21105/joss.03201},
+  url       = {https://doi.org/10.21105/joss.03201},
+  year      = {2021},
+  publisher = {The Open Journal},
+  volume    = {6},
+  number    = {63},
+  pages     = {3201},
+  author    = {David J.M. Barnett and Ilja C.W. Arts and John Penders},
+  title     = {microViz: an R package for microbiome data visualization and statistics},
+  journal   = {Journal of Open Source Software}
+}
+@article{bolyen2019,
+  doi       = {10.1038/s41587-019-0209-9},
+  title     = {Reproducible, interactive, scalable and extensible microbiome data science using QIIME 2},
+  author    = {Bolyen, Evan and Rideout, Jai Ram and Dillon, MaThew R and Bokulich, Nicholas A and Abnet, Christian C and Al-Ghalith, Gabriel A and Alexander, Harriet and Alm, Eric J and Arumugam, Manimozhiyan and Asnicar, Francesco and others},
+  journal   = {Nature biotechnology},
+  volume    = {37},
+  number    = {8},
+  pages     = {852--857},
+  year      = {2019},
+  publisher = {Nature Publishing Group}
+}
+
+@article{liu2020,
+    author = {Liu, Chi and Cui, Yaoming and Li, Xiangzhen and Yao, Minjie},
+    title = "{microeco: an R package for data mining in microbial community ecology}",
+    journal = {FEMS Microbiology Ecology},
+    volume = {97},
+    number = {2},
+    pages = {fiaa255},
+    year = {2020},
+    month = {12},
+    abstract = "{A large amount of sequencing data is produced in microbial community ecology studies using the high-throughput sequencing technique, especially amplicon-sequencing-based community data. After conducting the initial bioinformatic analysis of amplicon sequencing data, performing the subsequent statistics and data mining based on the operational taxonomic unit and taxonomic assignment tables is still complicated and time-consuming. To address this problem, we present an integrated R package-'microeco' as an analysis pipeline for treating microbial community and environmental data. This package was developed based on the R6 class system and combines a series of commonly used and advanced approaches in microbial community ecology research. The package includes classes for data preprocessing, taxa abundance plotting, venn diagram, alpha diversity analysis, beta diversity analysis, differential abundance test and indicator taxon analysis, environmental data analysis, null model analysis, network analysis and functional analysis. Each class is designed to provide a set of approaches that can be easily accessible to users. Compared with other R packages in the microbial ecology field, the microeco package is fast, flexible and modularized to use and provides powerful and convenient tools for researchers. The microeco package can be installed from CRAN (The Comprehensive R Archive Network) or github (https://github.com/ChiLiubio/microeco).}",
+    issn = {0168-6496},
+    doi = {10.1093/femsec/fiaa255},
+    url = {https://doi.org/10.1093/femsec/fiaa255},
+    eprint = {https://academic.oup.com/femsec/article-pdf/97/2/fiaa255/36115026/fiaa255.pdf},
+}
[…]

[…]
