### 2023-12-11 #6038 @iimog
Update from my side: I went through the manuscript and the repository. My general impression is, that MiscMetabar is a mature R package with a lot of useful functionality. @adrientaudiere put a lot of effort into development, testing, and documentation. I found a couple of minor issues, that are linked to this thread. I'm happy to re-evaluate and tick off the remaining boxes, as soon as they are addressed.
