+++ vignettes/tree_visualization.Rmd
@@ -2,27 +2,42 @@
 title: "Tree visualization"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{tree_visualization}
+  %\VignetteIndexEntry{Tree visualization}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 bibliography: bibliography.bib
 ---
 
 ```{r}
+library("MiscMetabar")
+library("phangorn")
+df <- subset_taxa_pq(data_fungi, taxa_sums(data_fungi) > 9000)
+df_tree <- build_phytree_pq(df, nb_bootstrap = 5)
+data_fungi_tree <- merge_phyloseq(df, df_tree$ML$tree)
+```
+
+
+```{r}
 library("treeio")
 library("ggtree")
-ggtree(ps_tree@phy_tree, layout = "ellipse") + geom_tiplab()
+
+ggtree(data_fungi_tree@phy_tree, layout = "ellipse") + geom_tiplab()
 ggtree(as.treedata(df_tree$ML), layout = "slanted")
 ggdensitree(df_tree$ML_bs, alpha = .3, colour = "steelblue") +
   geom_tiplab(size = 3) + hexpand(.35)
 ggtree(as.treedata(df_tree$ML)) +
   geom_text(aes(x = branch, label = AA_subs, vjust = -.5), size = 1)
-tax_tab <- as_tibble(ps_tree@tax_table) %>%
-  mutate(.otu = as.character(.otu))
-p <- ggtree(as.treedata(ps_tree@phy_tree)) %<+% as_tibble(ps_tree@tax_table)
+tax_tab <- as_tibble(data_fungi_tree@tax_table) |>
+  dplyr::mutate(.otu = as.character(.otu))
+p <- ggtree(as.treedata(data_fungi_tree@phy_tree)) %<+%
+  as_tibble(data_fungi_tree@tax_table)
 p + geom_tippoint(aes(color = Class, shape = Phylum)) +
   geom_text(aes(label = Genus), hjust = -0.2, size = 2)
-ggtree(as.treedata(ps_tree@phy_tree), branch.length = "none") %<+%
-  as_tibble(ps_tree@tax_table) +
+ggtree(as.treedata(data_fungi_tree@phy_tree), branch.length = "none") %<+%
+  as_tibble(data_fungi_tree@tax_table) +
   geom_tippoint(aes(color = Class, shape = Phylum), size = 2)
+
+# removeClass("phylo", where = "package:tidytree")
 ```
+
+
