+++ vignettes/filter.Rmd
@@ -55,7 +55,8 @@ In some cases, we want to select only a given clade of taxon. One solution is to
 ```{r}
 path_db <- system.file("extdata",
   "100_sp_UNITE_sh_general_release_dynamic.fasta",
-  package = "MiscMetabar", mustWork = TRUE)
+  package = "MiscMetabar", mustWork = TRUE
+)
 
 suppressWarnings(blast_error_or_not <-
   try(system("blastn 2>&1", intern = TRUE), silent = TRUE))
