+++ R/controls.R
@@ -24,7 +24,7 @@ search_exact_seq_pq <- function(physeq, sequences) {
   for (i in seq_along(sequences)) {
     original <- sequences[[i]]
     rev <- reverse(sequences[[i]])
-    rev_comp <-  Biostrings::reverseComplement(sequences[[i]])
+    rev_comp <- Biostrings::reverseComplement(sequences[[i]])
     comp <- complement(sequences[[i]])
 
     original_count <- sum(grepl(original, physeq@refseq))
