+++ vignettes/bibliography.bib
@@ -313,6 +313,13 @@
   journal   = {Journal of Open Source Software}
 }
 
+@manual{smith2023,
+  title  = {phylosmith: Functions to help analyze data as phyloseq objects},
+  author = {Schuyler Smith},
+  year   = {2023},
+  note   = {R package version 1.0.7},
+  url    = {https://schuyler-smith.github.io/phylosmith/}
+}
  @misc{taberlet2012,
   doi       = {10.1002/(issn)2637-4943},
   title     = {Environmental dna},
