+++ vignettes/import_export_track.Rmd
@@ -8,11 +8,11 @@ vignette: >
 ---
 
 
-
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  message = FALSE
 )
 ```
 
