+++ DESCRIPTION
@@ -71,6 +71,7 @@ Suggests:
     testthat (>= 3.0.0),
     tibble,
     tidyr,
+    tidytree,
     treeio,
     treemapify,
     utils, 
