### 2023-12-11 adrientaudiere/MiscMetabar#48 @adrientaudiere
Oh yes, I confess I didn't read this section. Thank you for pointing out this problem. I have carefully read all issue templates and adapted them to R package.
