+++ R/data.R
@@ -29,3 +29,19 @@
 #' - *Diameter*: diameter of the trunk
 #' - *Time*: time since the dead of the tree
 "data_fungi_sp_known"
+
+#' This tutorial explore the dataset from Tengeler et al. (2020) available in the `mia` package.
+#' obtain using `mia::makePhyloseqFromTreeSE(Tengeler2020)` 
+#' 
+#' This is a phyloseq version of the Tengeler2020 dataset.
+#' 
+#' Tengeler2020 includes gut microbiota profiles of 27 persons with ADHD.
+#' A standard bioinformatic and statistical analysis done to demonstrate that altered microbial 
+#' composition could be a driver of altered brain structure and function and concomitant changes 
+#' in the animals behavior. This was investigated by colonizing young, male, 
+#' germ-free C57BL/6JOlaHsd mice with microbiota from individuals with and without ADHD.
+#'
+#' @format 
+#' A phyloseq object
+#' 
+"Tengeler2020_pq"
