+++ R/blast.R
@@ -48,7 +48,8 @@
 #' data(data_fungi)
 #' blastpath <- "...YOUR_PATH_TO_BLAST..."
 #' blast_to_phyloseq(data_fungi,
-#'   seq2search = "inst/extdata/ex.fasta",
+#'   seq2search = system.file("extdata", "ex.fasta",
+#'   package = "MiscMetabar", mustWork = TRUE)
 #'   blastpath = blastpath
 #' )
 #' }
