+++ .github/workflows/pkgdown.yaml
@@ -37,7 +37,7 @@ jobs:
         run: sudo apt-get install ncbi-blast+
 
       - name: Build site
-        run: pkgdown::build_site_github_pages(new_process = FALSE, install = TRUE)
+        run: pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE, clean = TRUE)
         shell: Rscript {0}
 
       - name: Deploy to GitHub pages 🚀
