+++ pkgdown.yml
@@ -10,7 +10,9 @@ articles:
   filter: filter.html
   import_export_track: import_export_track.html
   states_of_fields_in_R: states_of_fields_in_R.html
-last_built: 2023-11-09T20:01Z
+  tengeler: tengeler.html
+  tree_visualization: tree_visualization.html
+last_built: 2023-12-11T18:04Z
 urls:
   reference: https://adrientaudiere.github.io/MiscMetabar/reference
   article: https://adrientaudiere.github.io/MiscMetabar/articles
