### 2023-12-12 b99a717
document: better Readme and vignettes MiscMetabar
