### 2023-12-12 b724ede
test: add mumu to installation for test-coverage
