+++ tests/testthat/test_figures_taxo.R
@@ -351,37 +351,37 @@ test_that("tax_bar_pq work with data_fungi dataset", {
   )
 })
 
-test_that("add_funguild_info and plot_guild_pq work with data_fungi dataset", {
-  expect_s4_class(
-    df <-
-      add_funguild_info(
-        subset_taxa_pq(data_fungi, taxa_sums(data_fungi) > 5000),
-        taxLevels = c(
-          "Domain",
-          "Phylum",
-          "Class",
-          "Order",
-          "Family",
-          "Genus",
-          "Species"
-        )
-      ),
-    "phyloseq"
-  )
-  expect_error(df <-
-    add_funguild_info(
-      subset_taxa_pq(data_fungi, taxa_sums(data_fungi) > 5000),
-      taxLevels = c(
-        "PHYLLUUM",
-        "Phylum",
-        "Class",
-        "Order",
-        "Family"
-      )
-    ))
-  expect_s3_class(plot_guild_pq(df, clean_pq = TRUE), "ggplot")
-  expect_s3_class(plot_guild_pq(df, clean_pq = FALSE), "ggplot")
-})
+# test_that("add_funguild_info and plot_guild_pq work with data_fungi dataset", {
+#   expect_s4_class(
+#     df <-
+#       add_funguild_info(
+#         subset_taxa_pq(data_fungi, taxa_sums(data_fungi) > 5000),
+#         taxLevels = c(
+#           "Domain",
+#           "Phylum",
+#           "Class",
+#           "Order",
+#           "Family",
+#           "Genus",
+#           "Species"
+#         )
+#       ),
+#     "phyloseq"
+#   )
+#   expect_error(df <-
+#     add_funguild_info(
+#       subset_taxa_pq(data_fungi, taxa_sums(data_fungi) > 5000),
+#       taxLevels = c(
+#         "PHYLLUUM",
+#         "Phylum",
+#         "Class",
+#         "Order",
+#         "Family"
+#       )
+#     ))
+#   expect_s3_class(plot_guild_pq(df, clean_pq = TRUE), "ggplot")
+#   expect_s3_class(plot_guild_pq(df, clean_pq = FALSE), "ggplot")
+# })
 
 
 test_that("build_phytree_pq work with data_fungi dataset", {
