+++ vignettes/MiscMetabar.Rmd
@@ -12,7 +12,7 @@ vignette: >
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  message=FALSE
+  message = FALSE
 )
 ```
 
