+++ man/blast_to_phyloseq.Rd
@@ -82,7 +82,8 @@ the blast table
 data(data_fungi)
 blastpath <- "...YOUR_PATH_TO_BLAST..."
 blast_to_phyloseq(data_fungi,
-  seq2search = "extdata/ex.fasta",
+  seq2search = system.file("extdata", "ex.fasta",
+  package = "MiscMetabar", mustWork = TRUE)
   blastpath = blastpath
 )
 }
