### 2023-12-12 c81431e
bugfix: retry to install mumu
