+++ .github/workflows/test-coverage.yaml
@@ -32,7 +32,7 @@ jobs:
         run: sudo apt-get install ncbi-blast+
 
       - name: Install mumu
-        run: >
+        run: |
          git clone https://github.com/frederic-mahe/mumu.git \
          && cd ./mumu/\
          && make\
