### 2023-12-12 adrientaudiere/MiscMetabar#42 @adrientaudiere
The new version of the paper resolves all the problems.
