+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   states_of_fields_in_R: states_of_fields_in_R.html
   tengeler: tengeler.html
   tree_visualization: tree_visualization.html
-last_built: 2023-12-11T18:04Z
+last_built: 2023-12-12T08:59Z
 urls:
   reference: https://adrientaudiere.github.io/MiscMetabar/reference
   article: https://adrientaudiere.github.io/MiscMetabar/articles
