### 2023-12-12 adrientaudiere/MiscMetabar#45 @adrientaudiere
Solved in pkgdown documentation for v0.51.
