+++ .github/workflows/test-coverage.yaml
@@ -34,9 +34,9 @@ jobs:
       - name: Install mumu
         run: |
          git clone https://github.com/frederic-mahe/mumu.git \
-         && cd ./mumu/\
-         && make\
-         && make check\
+         && cd ./mumu/ \
+         && make \
+         && make check \
          && sudo make install 
 
       - name: Test coverage
