### 2023-12-12 c2e7e90
style: improve page references from pkgdown site
