+++ paper/bibliography.bib
@@ -221,7 +221,8 @@
   number    = {1},
   pages     = {4899},
   year      = {2023},
-  publisher = {Nature Publishing Group UK London}
+  publisher = {Nature Publishing Group UK London},
+  doi       = {10.1038/s41467-023-39876-6}
 }
 
 
