### 2023-12-12 f342946
document: News in v0.52
