+++ .github/workflows/R-CMD-check.yaml
@@ -41,14 +41,8 @@ jobs:
         with:
           extra-packages: any::rcmdcheck
           needs: check
-      
-      - name: Install vsearch
-        run: sudo apt-get install vsearch
-
-      - name: Install blastn
-        run: sudo apt-get install ncbi-blast+
 
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
-          args: 'c("--no-tests", "--as-cran", "--no-vignettes", "--no-build-vignettes", "--timing")'
+          args: 'c("--no-tests", "--no-examples", "--no-vignettes", "--no-build-vignettes")'
