### 2023-12-12 adrientaudiere/MiscMetabar#38 @iimog
After merging the pr #43, these instructions are no longer required in the README:
```R
if (!require("BiocManager", quietly = TRUE))
  install.packages("BiocManager")

BiocManager::install("dada2")
BiocManager::install("phyloseq")
```
It does not hurt to run these commands beforehand, it's just not necessary anymore.
