+++ .github/workflows/R-CMD-check.yaml
@@ -45,4 +45,7 @@ jobs:
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
-          args: 'c("--no-tests", "--no-examples", "--no-vignettes", "--no-build-vignettes", "--no-manual")'
+          #Character vector of arguments to pass to ⁠R CMD check⁠. 
+          args: 'c("--as-cran",  , "--no-tests", "--no-vignettes", "--no-build-vignettes", "--ignore-vignettes")'
+          #Character vector of arguments to pass to ⁠R CMD build.
+          build_args: 'c("--no-build-vignettes", "--no-manual")'
