### 2023-12-12 #6038 @adrientaudiere
Thanks a lot @iimog @tkchafin for your review. It drastically improves the code and helps a lot improving the paper and the package.
