### 2023-12-12 017d373
documentation: improve README
