+++ man/count_seq.Rd
@@ -22,7 +22,8 @@ Use grep to count the number of line with only one '+' (fastq, fastq.gz)
 or lines starting with a '>' (fasta) to count sequences.
 }
 \examples{
-count_seq(file_path = system.file("extdata", "ex.fasta", package = "MiscMetabar", mustWork = TRUE))
+count_seq(file_path = system.file("extdata", "ex.fasta",
+          package = "MiscMetabar", mustWork = TRUE))
 count_seq(folder_path = "extdata/", pattern = "*.fasta")
 }
 \author{
