+++ paper/bibliography.bib
@@ -222,7 +222,8 @@
   pages     = {4899},
   year      = {2023},
   publisher = {Nature Publishing Group UK London},
-  doi       = {10.1038/s41467-023-39876-6}
+  doi       = {10.1038/s41467-023-39876-6},
+  url       = {https://doi.org/10.1038/s41467-023-39876-6}
 }
 
 
