+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   states_of_fields_in_R: states_of_fields_in_R.html
   tengeler: tengeler.html
   tree_visualization: tree_visualization.html
-last_built: 2023-12-12T14:15Z
+last_built: 2023-12-12T14:31Z
 urls:
   reference: https://adrientaudiere.github.io/MiscMetabar/reference
   article: https://adrientaudiere.github.io/MiscMetabar/articles
