### 2023-12-12 52cf429
feature: add 2 functions
