### 2023-12-12 bdb5d6a
style and test: in direction of v0.52
