+++ tests/testthat/test_figures_alpha_div.R
@@ -155,3 +155,4 @@ test_that("accu_samp_threshold works with data_fungi dataset", {
     "ggplot"
   )
 })
+
