+++ paper/paper.bib
@@ -128,7 +128,9 @@
   number    = {1},
   pages     = {4899},
   year      = {2023},
-  publisher = {Nature Publishing Group UK London}
+  publisher = {Nature Publishing Group UK London},
+  doi      = {10.21203/rs.3.rs-2011054/v1},
+  url      = {https://app.dimensions.ai/details/publication/pub.1151239188}
 }
 
 @software{mclaren2020,
