+++ man/accu_plot.Rd
@@ -67,7 +67,9 @@ return a base plot.
 \examples{
 data("GlobalPatterns", package = "phyloseq")
 GP <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == "Archaea")
+GP <- rarefy_even_depth(subset_samples_pq(GP, sample_sums(GP) > 3000))
 p <- accu_plot(GP, "SampleType", add_nb_seq = TRUE, by.fact = TRUE, step = 10)
+p <- accu_plot(GP, "SampleType", add_nb_seq = TRUE, step = 10)
 
 p + theme(legend.position = "none")
 
