+++ NEWS.md
@@ -1,4 +1,6 @@
-# MiscMetabar 0.52 (in development) 
+# MiscMetabar 0.53 (in development) 
+
+# MiscMetabar 0.52 
 
 - Add function `ggbetween_pq()` to facilitate comparison of hill number using the power of `ggstatsplot::ggbetweenstats()`
 - Add function `plot_SCBD_pq()` to plot species contributions to beta diversity (SCBD) of samples
