+++ man/accu_samp_threshold.Rd
@@ -22,6 +22,7 @@ to obtain \code{threshold} proportion of the ASV
 \examples{
 data("GlobalPatterns", package = "phyloseq")
 GP <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == "Archaea")
+GP <- rarefy_even_depth(subset_samples_pq(GP, sample_sums(GP) > 3000))
 p <- accu_plot(GP, "SampleType", add_nb_seq = TRUE, by.fact = TRUE, step = 10)
 
 val_threshold <- accu_samp_threshold(p)
