### 2023-12-13 f62fbff
test: fix a bug in test
