### 2023-12-13 #6038 @adrientaudiere
Hi @kellyrowland,
This is done for v. 0.52 : [zenodo](https://doi.org/10.5281/zenodo.10370781)
doi : 10.5281/zenodo.10370781
