+++ vignettes/beta-div.Rmd
@@ -156,7 +156,7 @@ sudo apt-get install libgsl-dev libmpfr-dev
 ### Using Deseq2 package
 
 ```{r}
-data("GlobalPatterns")
+data("GlobalPatterns", package = "phyloseq")
 GP <- subset_samples(
   GlobalPatterns,
   GlobalPatterns@sam_data$SampleType %in% c("Soil", "Skin")
