+++ paper/paper_old.md
@@ -26,16 +26,16 @@ Describing communities of living organisms increasingly relies on massive DNA se
 
 Biological studies, especially in ecology, health sciences and taxonomy, need to describe the biological composition of samples. During the last twenty years, (i) the development of DNA sequencing, (ii) reference databases, (iii) high-throughput sequencing (HTS), and (iv) bioinformatics resources have allowed the description of biological communities through metabarcoding. Metabarcoding involves the sequencing of millions (*meta*-) of short regions of specific DNA (*-barcoding*, @valentini2009) often from environmental samples (eDNA, @taberlet2012) such as human stomach contents, lake water, soil and air.
 
-The analysis of metabarcoding studies is difficult due to the complexity of the datasets (millions of sequences of variable quality, including chimeras and sequencing errors), often coupled with tricky sampling schemes and a wide variety of questions. Several plateforms (see Table 1 in @tedersoo2022 for a list) such as QIIME2 (@bolyen2019), mothur (@schloss2020), and Galaxy (@jalili2020) allow complete analysis from raw fastq sequences to statistical analysis and visualization. However, the R ecosystem (@rcran), especially through the [bioconductor](http://bioconductor.org/) project, is very rich (see the "State of the field in R" section) and more flexible than these platforms.
+The analysis of metabarcoding studies is difficult due to the complexity of the datasets (millions of sequences of variable quality, including chimeras and sequencing errors), often coupled with tricky sampling schemes and a wide variety of questions. Several plateforms (see Table 1 in @tedersoo2022 for a list) such as QIIME2 (@bolyen2019), mothur (@schloss2020), and Galaxy (@jalili2020) allow complete analysis from raw fastq sequences to statistical analysis and visualization. However, the R ecosystem (@rcran), especially through the [bioconductor](https://bioconductor.org/) project, is very rich (see the "State of the field in R" section) and more flexible than these platforms.
 
 `MiscMetabar` aims to facilitate the **description**, **transformation**, **exploration** and **reproducibility** of metabarcoding analysis using R. The development of `MiscMetabar` relies heavily on the R packages `dada2`, `phyloseq` and `targets`.
 
 
 # State of the Field in R
 
-The metabarcoding ecosystem in the R language is mature, well-constructed, and relies on a very active community in both the [bioconductor](https://www.bioconductor.org/) and [cran](https://cran.r-project.org/) projects. The [bioconductor](https://www.bioconductor.org/) even creates specific task views in [Metagenomics](http://bioconductor.org/packages/release/BiocViews.html#___Metagenomics) and [Microbiome](http://bioconductor.org/packages/release/BiocViews.html#___Microbiome).
+The metabarcoding ecosystem in the R language is mature, well-constructed, and relies on a very active community in both the [bioconductor](https://www.bioconductor.org/) and [cran](https://cran.r-project.org/) projects. The [bioconductor](https://www.bioconductor.org/) even creates specific task views in [Metagenomics](https://bioconductor.org/packages/release/BiocViews.html#___Metagenomics) and [Microbiome](https://bioconductor.org/packages/release/BiocViews.html#___Microbiome).
 
[…]
