+++ paper/paper.bib
@@ -152,7 +152,7 @@
   volume  = {8},
   number  = {4},
   year    = {2013},
-  url     = {http://dx.plos.org/10.1371/journal.pone.0061217}
+  url     = {https://dx.plos.org/10.1371/journal.pone.0061217}
 }
 
  @article{ondov2011,
