+++ R/miscellanous.R
@@ -269,9 +269,6 @@ funky_color <-
   )
 ################################################################################
 
-
-################################################################################
-
 ################################################################################
 #' Subsample a fastq file copying the n_seq first sequences in a given folder
 #'
