### 2023-12-15 #6038 @Kevin-Mattheus-Moerman
@adrientaudiere As AEiC I will now help process this work for acceptance. I have checked this review, the paper, the repository, and the archive link. Most seems in order, I do however have the below points that need your attention.

On the archive:
- [ ] Note, it is not a requirement, but we recommend that you add the ORCID link for the author listed on the ZENODO archive.

On the paper:
- [ ] The paper title currently contains: `MiscMetabar : an R packages...`, this should probably be edited to: `MiscMetabar: an R package...`, or else `MiscMetabar: R packages...` if it does refer to multiple. Note I also removed the space between the project name and `:`.

Missing DOIs:

- [ ] [10.1093/nar/gkaa434](https://doi.org/10.1093/nar/gkaa434) for:
```
Jalili, V., Afgan, E., Gu, Q., Clements, D., Blankenberg, D., Goecks, J., Taylor, J., &
Nekrutenko, A. (2020). The galaxy platform for accessible, reproducible and collaborative
biomedical analyses: 2020 update. Nucleic Acids Research, 48(W1), W395–W402.
```
- [ ] [10.1186/1471-2105-12-385](https://doi.org/10.1186/1471-2105-12-385) for:
```
Ondov, B.D., Bergman, N.H. & Phillippy, A.M. Interactive metagenomic visualization in a Web browser. BMC Bioinformatics 12, 385 (2011).
```

- [ ] [10.1128/spectrum.01978-23](https://doi.org/10.1128/spectrum.01978-23) for:
```
Vieira, F. R., Di Tomassi, I., O’Connor, E., Bull, C. T., Pecchia, J. A., & Hockett, K. L.
(2023). Manipulating agaricus bisporus developmental patterns by passaging microbial
communities in complex substrates. Microbiology Spectrum, e01978–23.
```

- [ ] [10.1093/procel/pwad024](https://doi.org/10.1093/procel/pwad024) for:
```
Wen, T., Niu, G., Chen, T., Shen, Q., Yuan, J., & Liu, Y.-X. (2023). The best practice for
microbiome analysis using r. Protein & Cell, pwad024
```
