### 2023-12-19 #6038 @adrientaudiere
Done. I also put the R of reproducibility in lowercase, like in the paper title.

> MiscMetabar: an R package to facilitate visualization and reproducibility in metabarcoding analysis
