### 2023-12-19 f3865e0
Create citation.cff
