### 2023-12-19 ffe1586
Doc: Better documentation and function are_modality_even_depth
