+++ .github/workflows/test-coverage.yaml
@@ -1,9 +1,8 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
-  push:
-    branches: 
-      - master
+  pull_request:
+    branches: [main, master]
 
 name: test-coverage
 
