### 2023-12-19 2e0274f
Doc: better documentation and README
