+++ man/upset_test_pq.Rd
@@ -33,7 +33,7 @@ will not count in the venn diagram}
 if FALSE, NA values are set to "NA"}
 
 \item{numeric_fonction}{(default : sum) the function for numeric vector
-usefull only for complex plot (see examples)}
+useful only for complex plot (see examples)}
 
 \item{...}{other arguments passed on to the \code{\link[ComplexUpset:upset_test]{ComplexUpset::upset_test()}}}
 }
