+++ R/plot_functions.R
@@ -1031,7 +1031,7 @@ ggvenn_pq <- function(physeq = NULL,
   nb_samples <- table(physeq@sam_data[[fact]])
 
   if (rarefy_after_merging) {
-    physeq <- speedyseq::merge_samples2(physeq, fact)
+    physeq <- merge_samples2(physeq, fact)
     physeq <- rarefy_even_depth(physeq)
     physeq <- clean_pq(physeq)
   }
@@ -1409,12 +1409,12 @@ hill_pq <-
 #'
 #' @description
 #' `r lifecycle::badge("experimental")`
-#' 
-#' Note that contrary to [hill_pq()], this function does not take into 
+#'
+#' Note that contrary to [hill_pq()], this function does not take into
 #' account for difference in the number of sequences per samples/modalities.
-#' You may use rarefy_by_sample = TRUE if the mean number of sequences per 
-#' samples differs among modalities. 
-#' 
+#' You may use rarefy_by_sample = TRUE if the mean number of sequences per
+#' samples differs among modalities.
+#'
 #' Basically a wrapper of function [ggstatsplot::ggbetweenstats()] for
 #' object of class phyloseq
 #' @inheritParams clean_pq
@@ -1422,7 +1422,7 @@ hill_pq <-
 #'   the `sam_data` slot of the physeq object.
 #' @param one_plot (logical, default FALSE) If TRUE, return a unique
 #'   plot with the three plot inside using the patchwork package.
-#' @param rarefy_by_sample (logical, default FALSE) If TRUE, rarefy 
+#' @param rarefy_by_sample (logical, default FALSE) If TRUE, rarefy
 #'   samples using [phyloseq::rarefy_even_depth()] function
 #' @param ... Other arguments passed on to [ggstatsplot::ggbetweenstats()] function.
 
@@ -1437,6 +1437,7 @@ hill_pq <-
 #'
 #' @export
 #' @examples
+#' data(data_fungi)
 #' p <- ggbetween_pq(data_fungi, variable = "Time", p.adjust.method = "BH")
 #' p[[1]]
 #' ggbetween_pq(data_fungi, variable = "Height", one_plot = TRUE)
@@ -1453,10 +1454,11 @@ ggbetween_pq <- function(physeq, variable, one_plot = FALSE, rarefy_by_sample = 
     physeq <- clean_pq(rarefy_even_depth(physeq))
   }
 
-  if(are_modality_even_depth(physeq, variable)$p.value < 0.05){
-    warning(paste0("The mean number of sequences per samples vary across modalities of the variable '", 
-    variable, 
-    "' You should use rarefy_by_sample = TRUE or try hill_pq() with correction_for_sample_size = TRUE"
+  if (are_modality_even_depth(physeq, variable)$p.value < 0.05) {
+    warning(paste0(
+      "The mean number of sequences per samples vary across modalities of the variable '",
+      variable,
+      "' You should use rarefy_by_sample = TRUE or try hill_pq() with correction_for_sample_size = TRUE"
     ))
   }
 
@@ -1818,7 +1820,7 @@ heat_tree_pq <- function(physeq, taxonomic_level = NULL, ...) {
 #'   If left to NULL use the `left_name` and `right_name` parameter as modality.
 #' @param merge_sample_by (default: NULL) if not `NULL` samples of
 #'   physeq are merged using the vector set by `merge_sample_by`. This
[…]
