+++ R/krona.R
@@ -100,7 +100,7 @@ krona <-
 #' ([installation instruction](https://github.com/marbl/Krona/wiki/Installing)).
 #' Function merge_krona allows merging multiple html files in one interactive
 #' krona file
-#' 
+#'
 #' Note that you need to use the name args in `krona()` function before `merge_krona()`
 #' in order to give good name to each krona pie in the output.
 #' @param files (required) path to html files to merged
@@ -108,11 +108,11 @@ krona <-
 #'
 #' @examples
 #' \dontrun{
-#'   data("GlobalPatterns", package = "phyloseq")
-#'   GA <- subset_taxa(GlobalPatterns, Phylum == "Acidobacteria")
-#'   krona(GA, "Number.of.sequences.html", name="Nb_seq_GP_acidobacteria")
-#'   krona(GA, "Number.of.ASVs.html", nb_seq = FALSE, name="Nb_asv_GP_acidobacteria")
-#'   merge_krona(c("Number.of.sequences.html", "Number.of.ASVs.html"))
+#' data("GlobalPatterns", package = "phyloseq")
+#' GA <- subset_taxa(GlobalPatterns, Phylum == "Acidobacteria")
+#' krona(GA, "Number.of.sequences.html", name = "Nb_seq_GP_acidobacteria")
+#' krona(GA, "Number.of.ASVs.html", nb_seq = FALSE, name = "Nb_asv_GP_acidobacteria")
+#' merge_krona(c("Number.of.sequences.html", "Number.of.ASVs.html"))
 #' }
 #' @return A html file
 #' @seealso \code{\link{krona}}
