+++ man/upset_pq.Rd
@@ -154,7 +154,7 @@ data_fungi2@sam_data[["Height__Time_0"]] <-
   paste0(data_fungi2@sam_data[["Height"]], "__", data_fungi2@sam_data[["Time_0"]])
 data_fungi2@sam_data[["Height__Time_0"]][grepl("NA", data_fungi2@sam_data[["Height__Time_0"]])] <-
   NA
-upset_pq(data_fungi2, fact = "Height__Time_0", width_ratio = 0.2)
+upset_pq(data_fungi2, fact = "Height__Time_0", width_ratio = 0.2, min_size = 2)
 }
 \seealso{
 \code{\link[=ggvenn_pq]{ggvenn_pq()}}

[…]
