+++ R/beta_div_test.R
@@ -161,7 +161,7 @@ adonis_pq <- function(physeq,
       )
   } else {
     phy_dist <-
-      paste0('phyloseq:::distance(physeq, method="', dist_method, '")')
+      paste0('phyloseq::distance(physeq, method="', dist_method, '")')
   }
 
   .formula <- stats::reformulate(formula, response = phy_dist)
@@ -231,6 +231,7 @@ adonis_pq <- function(physeq,
 #' @export
 #' @seealso [plot_LCBD_pq], [adespatial::beta.div()]
 #' @examples
+#' data(data_fungi_sp_known)
 #' res <- LCBD_pq(data_fungi_sp_known, nperm = 50)
 #' str(res)
 #' length(res$LCBD)
