### 2023-12-22 cf08f48
format: improve code quality
