+++ tests/testthat/test_figures_taxo.R
@@ -266,7 +266,7 @@ test_that("rigdes_pq work with data_fungi dataset", {
   expect_s3_class(
     ridges_pq(clean_pq(
       subset_taxa(data_fungi_sp_known, Phylum == "Basidiomycota")
-    ), alpha = 0.6, scale = 0.9),
+    ), "Time", alpha = 0.6, scale = 0.9),
     "ggplot"
   )
   expect_s3_class(
@@ -274,6 +274,7 @@ test_that("rigdes_pq work with data_fungi dataset", {
       clean_pq(subset_taxa(
         data_fungi_sp_known, Phylum == "Basidiomycota"
       )),
+      "Time",
       jittered_points = TRUE,
       position = ggridges::position_points_jitter(width = 0.05, height = 0),
       point_shape = "|",
@@ -287,8 +288,7 @@ test_that("rigdes_pq work with data_fungi dataset", {
   expect_error(
     ridges_pq(clean_pq(
       subset_taxa(data_fungi_sp_known, Phylum == "Basidiomycota")
-    )),
-    "ggplot"
+    ))
   )
 })
 
