+++ R/beta_div_test.R
@@ -11,8 +11,7 @@
 #'   Need to be in \code{physeq@sam_data}. This should be a factor
 #'   with two or more levels.
 #' @param merge_sample_by a vector to determine
-#'   which samples to merge using the
-#'   \code{\link[speedyseq]{merge_samples2}} function.
+#'   which samples to merge using [merge_samples2()] function.
 #'   Need to be in \code{physeq@sam_data}
 #' @param nperm (int) The number of permutations to perform.
 #' @param return_plot (logical) Do we return only the result
@@ -36,14 +35,15 @@
 #'   Please cite `phyloseqGraphTest` package.
 #' @export
 
-graph_test_pq <- function(physeq,
-                          fact,
-                          merge_sample_by = NULL,
-                          nperm = 999,
-                          return_plot = TRUE,
-                          title = "Graph Test",
-                          na_remove = FALSE,
-                          ...) {
+graph_test_pq <- function(
+        physeq,
+        fact,
+        merge_sample_by = NULL,
+        nperm = 999,
+        return_plot = TRUE,
+        title = "Graph Test",
+        na_remove = FALSE,
+        ...) {
   verify_pq(physeq)
 
   if (!is.null(merge_sample_by)) {
@@ -135,14 +135,15 @@ graph_test_pq <- function(physeq,
 #'   Please make a reference to `vegan::adonis2()` if you
 #'   use this function.
 
-adonis_pq <- function(physeq,
-                      formula,
-                      dist_method = "bray",
-                      merge_sample_by = NULL,
-                      na_remove = FALSE,
-                      correction_for_sample_size = FALSE,
-                      rarefy_nb_seqs = FALSE,
-                      ...) {
+adonis_pq <- function(
+        physeq,
+        formula,
+        dist_method = "bray",
+        merge_sample_by = NULL,
+        na_remove = FALSE,
+        correction_for_sample_size = FALSE,
+        rarefy_nb_seqs = FALSE,
+        ...) {
   physeq <- clean_pq(
     physeq,
     force_taxa_as_columns = TRUE,
@@ -242,9 +243,10 @@ adonis_pq <- function(physeq,
 #' This function is mainly a wrapper of the work of others.
 #'   Please make a reference to `adespatial::beta.div()` if you
 #'   use this function.
-LCBD_pq <- function(physeq,
-                    p_adjust_method = "BH",
-                    ...) {
+LCBD_pq <- function(
+        physeq,
+        p_adjust_method = "BH",
+        ...) {
   physeq <- clean_pq(
     physeq,
     force_taxa_as_columns = TRUE,
@@ -315,12 +317,13 @@ LCBD_pq <- function(physeq,
 #' This function is mainly a wrapper of the work of others.
 #'   Please make a reference to `vegan::beta.div()` if you
 #'   use this function.
-plot_LCBD_pq <- function(physeq,
-                         p_adjust_method = "BH",
-                         pval = 0.05,
-                         sam_variables = NULL,
-                         only_plot_significant = TRUE,
-                         ...) {
+plot_LCBD_pq <- function(
+        physeq,
+        p_adjust_method = "BH",
+        pval = 0.05,
+        sam_variables = NULL,
+        only_plot_significant = TRUE,
+        ...) {
   resBeta <- LCBD_pq(physeq,
     p_adjust_method = p_adjust_method,
     ...
@@ -438,11 +441,12 @@ plot_LCBD_pq <- function(physeq,
 #' This function is mainly a wrapper of the work of others.
 #'   Please make a reference to `vegan::beta.div()` if you
 #'   use this function.
-plot_SCBD_pq <- function(physeq,
[…]
