+++ vignettes/Reclustering.Rmd
@@ -51,7 +51,6 @@ summary_plot_pq(otu)
 Another post-clustering transformation method is implemented in `lulu_pq()`, which uses @froslev2017's method for curation of DNA amplicon data. The aim is more to clean non-biological information than to make explicitly less clusters.
 
 ```{r, results="hide"}
-library("lulu")
 data(data_fungi_sp_known)
 lulu_res <- lulu_pq(data_fungi_sp_known)
 ```
