+++ man/iNEXT_pq.Rd
@@ -38,11 +38,11 @@ GPsubset <- subset_taxa(
   GPsubset,
   rowSums(is.na(GPsubset@tax_table)) == 0
 )
-GPsubset@sam_data$human <- match(GPsubset@sam_data$SampleType,
-   c("Skin", "Feces", "Tong"))
+GPsubset@sam_data$human <- GPsubset@sam_data$SampleType \%in\%
+ c("Skin", "Feces", "Tong")
 res_iNEXT <- iNEXT_pq(
   GPsubset,
-  merge_sample_by = ,
+  merge_sample_by = human,
   q = 1,
   datatype = "abundance",
   nboot = 2
