### 2023-12-26 ce50c94
Update plot_functions.R
