### 2023-12-27 71c847d
Update beta_div_test.R
