+++ R/beta_div_test.R
@@ -489,6 +489,7 @@ plot_SCBD_pq <- function(
 #' @return A ggplot object
 #' @export
 #' @examples
+#' data(data_fungi) 
 #' data_fungi_ab <- subset_taxa_pq(data_fungi, taxa_sums(data_fungi) > 10000)
 #' multipatt_pq(subset_samples(data_fungi_ab, !is.na(Time)), fact = "Time")
 #' multipatt_pq(subset_samples(data_fungi_ab, !is.na(Time)),
