### 2023-12-28 0f33b11
Minor bugs fix to pass R CMD CHECK
