+++ R/plot_functions.R
@@ -21,11 +21,10 @@
 #' @seealso [phyloseq::mt()]
 
 plot_mt <-
-  function(
-        mt = NULL,
-        alpha = 0.05,
-        color_tax = "Class",
-        taxa = "Species") {
+  function(mt = NULL,
+           alpha = 0.05,
+           color_tax = "Class",
+           taxa = "Species") {
     d <- mt[mt$plower < alpha, ]
     d$tax_col <- factor(as.character(d[, color_tax]))
     d$tax_col[is.na(d$tax_col)] <- "unidentified"
@@ -99,19 +98,18 @@ plot_mt <-
 #' @author Adrien Taudière
 #' @seealso \code{\link[vegan]{specaccum}} [accu_samp_threshold()]
 accu_plot <-
-  function(
-        physeq,
-        fact = NULL,
-        add_nb_seq = TRUE,
-        step = NULL,
-        by.fact = FALSE,
-        ci_col = NULL,
-        col = NULL,
-        lwd = 3,
-        leg = TRUE,
-        print_sam_names = FALSE,
-        ci = 2,
-        ...) {
+  function(physeq,
+           fact = NULL,
+           add_nb_seq = TRUE,
+           step = NULL,
+           by.fact = FALSE,
+           ci_col = NULL,
+           col = NULL,
+           lwd = 3,
+           leg = TRUE,
+           print_sam_names = FALSE,
+           ci = 2,
+           ...) {
     if (!inherits(physeq, "phyloseq")) {
       stop("physeq must be a phyloseq object")
     }
@@ -374,21 +372,20 @@ accu_samp_threshold <- function(res_accuplot, threshold = 0.95) {
 #' @seealso \code{\link[circlize]{circos.par}}
 
 circle_pq <-
-  function(
-        physeq = NULL,
-        fact = NULL,
-        taxa = "Order",
-        nproc = 1,
-        add_nb_seq = TRUE,
-        rarefy = FALSE,
-        min_prop_tax = 0.01,
-        min_prop_mod = 0.1,
-        gap_degree = NULL,
-        start_degree = NULL,
-        row_col = NULL,
-        grid_col = NULL,
-        log10trans = FALSE,
-        ...) {
+  function(physeq = NULL,
+           fact = NULL,
+           taxa = "Order",
+           nproc = 1,
+           add_nb_seq = TRUE,
+           rarefy = FALSE,
+           min_prop_tax = 0.01,
+           min_prop_mod = 0.1,
+           gap_degree = NULL,
+           start_degree = NULL,
+           row_col = NULL,
+           grid_col = NULL,
+           log10trans = FALSE,
+           ...) {
     if (!inherits(physeq, "phyloseq")) {
       stop("physeq must be an object of class 'phyloseq'")
     }
@@ -573,16 +570,15 @@ circle_pq <-
 #' @seealso \code{\link[networkD3]{sankeyNetwork}}
 
 sankey_pq <-
-  function(
-        physeq = NULL,
-        fact = NULL,
-        taxa = 1:4,
-        add_nb_seq = FALSE,
-        min_prop_tax = 0,
-        tax2remove = NULL,
-        units = NULL,
-        symbol2sub = c("\\.", "-"),
-        ...) {
+  function(physeq = NULL,
+           fact = NULL,
+           taxa = 1:4,
+           add_nb_seq = FALSE,
+           min_prop_tax = 0,
+           tax2remove = NULL,
+           units = NULL,
+           symbol2sub = c("\\.", "-"),
+           ...) {
     if (!inherits(physeq, "phyloseq")) {
       stop("physeq must be an object of class 'phyloseq'")
     }
[…]
