### 2023-12-28 6ba6d93
Minor bugs fix to pass R CMD CHECK
