### 2023-12-28 d2d17ec
fix: put dontrun un iNEXT example
